Ensure SimulationRunnerIterator saves states when tearing down
Modified the SimulationRunnerIterator to ensure that the state of sub-simulation runners is saved before they are torn down. This means that setting the equilibration time at the CalcSet level will now work as expected (previously the state of the calculations was not saved, causing the analysis to fail).Also ensured that methods defined at the SimulationRunner level which are inappropriately called at the CalcSet level raise NotImplemented errors.